issue: 865066 Use one vma_extra.h file for both vmapoll and legacy sockets API
File vmapoll_extra.h was removed
